Replace `fast-glob` with `tinyglobby`
`tinyglobby` uses `picomatch` under the hood which default to
case-sensitive matching. Not sure what are the performance
implications of performing such matching on Unix systems, but in case
we want to preserve the distinction, an update in `tinyglobby` would be
needed to expose the underlying `picomatch` options or add a new
top-level one that'd control the `nocase` flag.